Update README.md
Let's go through each step in more detail:

Ask the user for the TikTok video URL they want to download without a watermark.

This step prompts the user to enter the URL of the TikTok video they want to download without a watermark. You can use a graphical user interface or a command line interface to get user input.

Parse the URL to extract the video ID.

In this step, you need to extract the video ID from the URL. TikTok video URLs typically look like https://www.tiktok.com/@username/video/1234567890123456789. The video ID in this URL is 1234567890123456789. You can use regular expressions or string manipulation to extract the video ID.

Construct the download URL using the video ID.

Once you have the video ID, you can construct the download URL by appending it to the base URL https://www.tiktok.com/node/share/video/. The resulting URL should look like https://www.tiktok.com/node/share/video/1234567890123456789.

Send an HTTP GET request to the download URL.

In this step, you need to send an HTTP GET request to the download URL to retrieve the video file. You can use a third-party library like requests to send the request and handle the response.

If the response is successful, save the video file to the desired location on the user's device.

Once you have the video file, you can save it to the desired location on the user's device. You can use the open function to create a new file and write the video content to it.

Remove the watermark from the video file using a watermark removal algorithm.

In this step, you need to remove the watermark from the video file. There are several approaches you can take to remove the watermark, including image processing techniques and machine learning algorithms. You can use a third-party library or implement your own algorithm to remove the watermark.

Save the new video file without the watermark to the desired location on the user's device.

Once you have removed the watermark, you can save the new video file without the watermark to the desired location on the user's device. You can use the same open function to create a new file and write the watermark-removed video content to it.

Notify the user that the download is complete and provide them with the location of the downloaded file